Enemy collision fix.

+++ deadline/src/main/java/org/game/message/EnemyRemoveMessage.java
@@ -1,4 +1,4 @@
 package org.game.message;
 
-public record EnemyRemoveMessage(int enemyId) implements Message {
+public record EnemyRemoveMessage(long enemyId) implements Message {
 }
